requirements.txt
alembic==1.9.1
anyio==3.6.2
apache-airflow==2.5.0
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-ftp==3.2.0
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.0
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==2.0.0
attrs==22.2.0
Babel==2.11.0
black==22.12.0
blinker==1.5
boto3==1.26.37
botocore==1.29.37
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
cron-descriptor==1.2.32
croniter==1.3.8
cryptography==38.0.4
Deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.1.0
flake8==6.0.0
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
glob2==0.7
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lazy-object-proxy==1.8.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mypy-extensions==0.4.3
numpy==1.24.0
packaging==22.0
pandas==1.5.2
pathspec==0.9.0
pendulum==2.1.2
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
pluggy==1.0.0
prison==0.2.1
psutil==5.9.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.13.0
PyJWT==2.6.0
pyrsistent==0.19.2
python-daemon==2.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-nvd3==0.15.0
python-slugify==7.0.0
pytz==2022.7
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
s3transfer==0.6.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.1.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.4.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.11.0
